tomcat/src/main/java/com/techcourse/db/InMemoryUserRepository.java
package com.techcourse.db;

import com.techcourse.model.User;
import java.util.Map;
import java.util.Optional;
import java.util.concurrent.ConcurrentHashMap;

public class InMemoryUserRepository {

    private static final Map<String, User> database = new ConcurrentHashMap<>();

    private InMemoryUserRepository() {
    }

    static {
        User user = new User(1L, "gugu", "password", "hkkang@woowahan.com");
        database.put(user.getAccount(), user);
    }

    public static void save(User user) {
        database.put(user.getAccount(), user);
    }

    public static Optional<User> findByAccount(String account) {
        return Optional.ofNullable(database.get(account));
    }

    public static Optional<User> findByAccountAndPassword(String account, String password) {
        return database.values().stream()
                .filter(user -> account.equals(user.getAccount()))
                .filter(user -> user.checkPassword(password))
                .findAny();
    }

    public static void deleteByAccount(String account) {
        database.remove(account);
    }
}
